+++ src/app/components/footer/footer-contact-details/styles.tsx
@@ -1,0 +1,30 @@
+export const footerContactInfoBoxStyles = {
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "flex-start",
+};
+
+export const footerContactDetailsBoxStyles = {
+    display: "flex",
+    marginTop: "20px",
+    marginBottom: "20px",
+};
+
+export const footerAddressDetailsBoxStyle = {
+    display: "flex",
+    flexDirection: "column",
+    marginLeft: "30px",
+};
+
+export const footerPersonalContactDetailsBoxStyle = {
+    display: "flex",
+    flexDirection: "column",
+};
+
+export const footerSocialMediaBoxStyle = {
+    marginBottom: "20px",
+    display: "flex",
+    justifyContent: "space-evenly",
+    width: "150px",
+};
+export const footerSocialMediaStyle = { color: "#949494" };
