+++ src/app/components/footer/styles.tsx
@@ -6,34 +6,3 @@ export const footerMainBoxStyles = {
     justifyContent: "space-between",
     padding: "30px",
 };
-
-export const footerContactInfoBoxStyles = {
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "flex-start",
-};
-
-export const footerContactDetailsBoxStyles = {
-    display: "flex",
-    marginTop: "20px",
-    marginBottom: "20px",
-};
-
-export const footerAddressDetailsBoxStyle = {
-    display: "flex",
-    flexDirection: "column",
-    marginLeft: "30px",
-};
-
-export const footerPersonalContactDetailsBoxStyle = {
-    display: "flex",
-    flexDirection: "column",
-};
-
-export const footerSocialMediaBoxStyle = {
-    marginBottom: "20px",
-    display: "flex",
-    justifyContent: "space-evenly",
-    width: "150px",
-};
-export const footerSocialMediaStyle = { color: "#949494" };
